tdf#138995 DOCX import: fix handling of textbox zorders
Regression from commit c6a2de20fc98b8559aa44e0383788eb9fbf50338 (oox:
import WPS shape with text as shape with textbox, 2014-06-18), the
problem was that a textbox's shape + textframe are internally 2 sdr
objects, so once GraphicZOrderHelper knows the current shape should be
on top of a shape+frame pair, it should suggest a larger ZOrder.

This is necessary till there is no setter version of
SwTextBoxHelper::getOrdNum(), which would allow import filters to ignore
this complexity, but that would be a larger change.

Change-Id: Ibbb1bcd9301eb369f25f211120f62be7c59b0fd2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109111
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>