tdf#138122: Fix vcl for iOS after 1249391074d55c10c865f3b4f95c8c861cad400b
Make vcl compile again for iOS and make the Collabora Office iOS app
work again when built against a master build of core.

For now, keep the old versions of the functions touched by
1249391074d55c10c865f3b4f95c8c861cad400b in vcl/ios/salios.cxx, and
move the modified versions to the new file vcl/osx/salmacos.cxx.

Keep the functions as they were except that ifdefs for MACOSX or IOS
are expanded. Keep the formatting as it was to make comparisons
easier. Thus add the new files to the clang-format exclusion list.

While at it, also move vcl/quartz/salgdiutils.cxx to vcl/osx as it is
compiled only for macOS anyway.

Change-Id: I990ef678f2263031d4a5af8cc547fffe185d17c4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109480
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>