devtools: handle the doc. model tree with attached obects
Instead of filling the document model tree with fill* methods,
refactor that to use objects that are attached (via get_id) to the
tree nodes directly. For this introduce DocumentModelTreeEntry
and subclasses, which implement what the current UNO object is
that is being shown for a tree node and a "fill" virtual method,
which is used to fill the child nodes of the tree when expanding
a tree node.

This makes the code much easier to work with and in addition it
makes it possible to have all the tree nodes to fill the content
on demand when expanded.

Change-Id: Id5a027e060af90e483181439568f17d0172d1b35
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109500
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>