tdf#116757 sw ChangesInMargin: add "Show Insertions in Margin"
layout mode to allow to show the original text inline,
i.e. only deletions, while insertions are showed in margin
(opposite of the Show Deletions in Margin mode).

Add a menu to the Show Changes icon on the
Changes toolbar with menu items for the 3 layout modes
"All Changes Inline", "Deletions in Margin" and
"Insertions in Margin" with the following behavior:

– Selecting a menu item turns on also the Show Changes mode
  to show the result immediately;

– Disabling Show Changes mode always show the final text, i.e.
  only insertions. Enabling Show Changes mode again shows the
  changes according to the last Show Changes mode selected in
  the menu.

– The new "Insertions in Margin" mode is intended to be only a
  run-time option, i.e. it's not possible to set it as a default
  Show Changes mode in Options, unlike "Deletions in Margin".

Change-Id: Ie4dc9fe3b4b201fa9976ef04877af649e242ab46
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109649
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>