tdf#135590 rotated paper sizes reported as portrait size they are not
since...

commit 11a6cd6e9377689d5dd8c5c931e40b534c1660c5
Date:   Fri Jun 29 11:36:03 2018 -0300

    Sets paper sizes listbox in print dialog

see as the above commit added some uses of doSloppyFit to vcl I imagine
the calls there want to be able to match rotated paper sizes, but in
the cases of tdf#135590 we don't want that behaviour because it doesn't
match what the user is presented with, the width and height are swapped.

So drop matching against swapped height/width by default, but let calls
added in 'Sets paper sizes listbox in print dialog' continue to match
rotated sizes.

Change-Id: I34aeddf12a7ca22234fbc6394487d3c8da7772ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109784
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>