tdf#138907 sw TitlePageDlg: Goto correct (inserted) page
Especially with inserted-in-the-middle pages, until the layout
was re-calculated, the new pages were not recognized.

This patch should bring this format option into pretty good shape.

I made a couple of decisions:
1.) Insert an additional content page when adding to end of document
    (otherwise the last index page turned into content).
2.) Ensure that index pages don't inherit page-renumbering

This CalcLayout could take a fair amount of time on a huge document.
I tested on a document that took over a minute to load.
It inserted pages in about 10 seconds. The experience seemed
reasonable to me given the known slowness of the document.

make UITest_writer_tests5 UITEST_TEST_NAME=\
titlePageWizard2.tdf138907.test_tdf138907

Change-Id: I6d7763ab10b46b5a459e9b62ed96b0194b2258b4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109830
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>