tdf#139940 oox: mark end before growing largest segment backwards
This fixes a regression in 7.1 (backported to 7.0.1) caused by
commit af2578f63e54a3dd844ced2f5dd4afc01159b6f5 for tdf#134183.

The largest segment has already been found, so we know where the
next segment starts. Now if we grow it backwards, that shouldn't
affect where it ends. So flag the end before growing the start,
otherwise we would have to recalculate the end again.

Change-Id: Iaae401d0438e533c7498be56f48fa153f7d3ff0a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110004
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>