tdf#124701 sc status bar: hide bad filtered row count
After loading the document, positioning the cell cursor
on filtered data, the status bar showed always the maximum
amount of the filtered data, e.g. "9 of 9 records found",
regardless of the real result, e.g. "6 of 9 records found",
because the ScTable::Query() function will not run until
we refresh the filters. As a workaround, we hide the
(often false) data in the meantime.

Change-Id: I904ca0c55e3afb276b11491c05430f985557f914
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110032
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>