sw: fix copying and deleting of table in section via API
This is a follow-up to 967955178baba4fd24ccd87e77ea8dd482fff305.

The first problem was that tables that contain protected cells refused
to be deleted, which caused an infinite loop in
SwXTextRange::DeleteAndInsert().

This also affected SwXTextTable::dispose(), and DDE tables.

Fix this by forcibly deleting even protected cells from UNO APIs;
protection is an UI feature.

The second problem was that pasting the table SwFEShell::Paste() would
set up aCpyPam so that it selects from the start node of the table to
the last text node in the table, excluding the table cell and table end
nodes, which caused:

DocumentContentOperationsManager.cxx:3548: An insufficient number of nodes were copied!

Fix this by setting up aCpyPam over the entire body section of the
clipboard document.

Change-Id: I8c6caee5e75260dff79f106efc11f9669feddc0c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110411
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>