lok: sc - suppress LOK editengine events for the calc input bar.
The calc input uses an EditEngine - but it emits events as if it is
part of the document - as well as emitting events as if it is a dialog
window. This causes horrible jumping artifacts to the top of the
document with word auto-completion and typing in cells a few page down.

Also remove obsolete tweak to turn off autocorrect for LOK, since
it was contradicted by: 2e7a5614934b0b26261f847c2e145e4ec58b53e2

Change-Id: I7352cd3b5ca6d18645cfd9e18959b0e44ec969d7
Signed-off-by: Michael Meeks <michael.meeks@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110422
Tested-by: Jenkins