oox: prefix VML shapetype ids with _x0000_t
Word 2013 refuses to even load a file that has a <v:shapetype
id="shapetype_75"> on some form control shape, reporting a misleading
error in a location far later when the top-level w:tbl that contains
the shape ends.

Using id="_x0000_t75" appears to work, so let's do that then.

Couldn't find any documentation on why this is so.

Change-Id: Ie22bb04244e24b00a1880544872ae8e281422405
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110493
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>