tdf#140236 WIN install libffi with Python
Regression from commit 6385d2c47ecea0080d92f5665892db3afedd5e0f
("python3: update to 3.8.4"). Previous Python versions included
the source for libffi on all platforms, but now just for MacOS.
So now LO must build that DLL on Windows, which was done in
commit f0232127d2e276aa5ea0830c30706e54b8a2030d ("libffi: build
DLL on Windows").

Since OpenSSL is installed in the program directory and the LO
Python directory, this currently works "accidentially". Not sure
the Python OpenSSL DLLs should be handled via some extra gbuild
package to separate both, but that can be done in an additional
change.

Change-Id: I4a42e39cc2f4434a9944aad32836f66ec2819931
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110537
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>