tdf#140252 Controls in Calc are always on layer 'controls'
The ctor of SdDrawLayer had set the name of the control layer to
'Controls' (uppercase), whereas the ctor of SdrLayerAdmin sets it to
'controls' (lowercase).
In result, for a control object in Calc the layer was not found and
therefore set to default layer ID 0. I have not investigated in detail
why it was not found, but simple change the name in Calc to use
'controls' too.
Second error was, that a control, which was inserted as default object
via ctrl+click, had not get 'controls' als layer.

Change-Id: Iaac349cda0b0dcf3f098ca3484308ac52a83aa38
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110636
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>