always optimize bitmap transform to translate+scale if possible (tdf#138068)
Commit 5ec9f5c4bedd5318 make OutputDevice::DrawTransformedBitmapEx()
always call DrawTransformBitmapExDirect() in preference to explicitly
converting to DrawBitmapEx() if the transformation wanted was at most
translate+scale, in the hopes that these days the backend implementations
work well enough. But it turns out only the Skia backend handles that
without loss of performance. So always do the conversion to DrawBitmapEx()
if possible. It's so simple that it possibly makes sense to always do
this, regardless of the backend.

Change-Id: I6eba68e672334c38433f53980f49400499f5d8e7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110716
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>