tdf#121245: also set ooLocale when language is set by command line
ooLocale is re-written in each invocation of langselect::prepareLocale
from Desktop::Init; previously an exception was when it was set by command
line. This special treatment was from initial '--language' implementation
in commit 805ff145a9e9ad58505c7ce636b12af032190ac5.
The problem here is this setting is the fallback for multiple code paths,
so many UI elements got language from old ooLocale data in that case.
This setting is re-written every time the program is started, so it should
be no harm in writing it also in command line case.

Alternatively we could try to fix all places to use configured system UI
language from MsLangId instead, or to create a backend layer for ooLocale
that itself uses that.

Change-Id: Idf7212901620bc158f704ba21747294634e48ae1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110776
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>