Qt don't fallthrough in inputMethodQuery
So my initiak code had a single case with fallthrough to the
default, but that is totally bogus for multiple cases. I honestly
have no idea, if that ever was correct. I also don't know, if
simply returning an empty QVariant is the correct error handling,
but the fallthroughs were definitly wrong.

Change-Id: I1880a960164eccbb6a5a3832c4ae9aaeda065d43
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/110946
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>