devtools: change the object inspector view to tabs
Until now we had only one tree view with top-level nodes for
services, interfaces, properties and methods. The problem with
this is that each one category has it's own distinct columns
(especially methods) so they can't fit well into just one tree
view. Services and interfaces categories are very simple so they
can be presented in a simpler way.

This change adds a tab-bar on top, for all the categories and each
one is now presented in its own tree view, which makes it possible
to modify the tree view to the specific category. Tree views are
currently copies, but they will be modified in the future into
what is more appropriate for the category.

Change-Id: Ib532df9bb7b9e0920fff57085d6ec4f1031b3ed8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111092
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>