tdf#121666 DOCX export: fix missing page break
When an empty paragraph has both page break and section
break, only the section break was exported to DOCX,
the page break was not, while MSO needs that to show
all page breaks.

The DOC and RTF exports were not changed by this fix
(not sure if the problem are present with them or not).

Co-authored-by: Tibor Nagy (NISZ)

Change-Id: I5deac0390666415e36a541c731dda4bc0a2ad6ad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111543
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>