remove unnecessary MouseButtonDown forward
I don't see how this scenario can happen, and if it does, why
we would want to send a MouseDown to the filter list. This hunk appears
in the "initial import". There is a "On OS/2 when clicking next to the
Popup menu..." comment on a different popup in this file so maybe there
was some workaround for some condition that no longer happens.

Change-Id: I15b4d57259047a20822685a00e1bf5b00aa21a3a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/111715
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>